[tracer] Normalize paths lexically and only in report_file_ops
The problem with std::filepath::weakly_canonical() is that it also resolves
symlinks which was never the intended behavior.  Also, it can throw if e.g. a
symlink is broken (it is similar to std::filepath::canonical() which is a
wrapper on realpath(3); see its ERRORS section) which happens in practice.

All the paths changed in this patch flow into report_file_ops() and are
ultimately consumed on the other side by the analyzer; the point of
normalization is beautification and providing a nice invariant to the analyzer,
shall it decide to compare the paths for some reason.